+++ lib/versionarte.dart
@@ -7,6 +7,6 @@ export 'src/providers/restful_versionarte_provider.dart';
 export 'src/providers/versionarte_provider.dart';
 
 // Models
-export 'src/models/store_versioning.dart';
+export 'src/models/distribution_manifest.dart';
 export 'src/models/versionarte_result.dart';
 export 'src/models/versionarte_status.dart';